+++ src/Microsoft.Health.Dicom.Metadata/Registration/Experimental/IBlobServiceBuilder.cs
@@ -3,7 +3,6 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using EnsureThat;
 using Microsoft.Health.Blob.Configs;
 
 namespace Microsoft.Extensions.DependencyInjection
@@ -13,18 +12,5 @@ namespace Microsoft.Extensions.DependencyInjection
         IServiceCollection Services { get; }
 
         BlobDataStoreConfiguration Configuration { get; }
-    }
-
-    internal sealed class BlobServiceBuilder : IBlobServiceBuilder
-    {
-        public IServiceCollection Services { get; }
-
-        public BlobDataStoreConfiguration Configuration { get; }
-
-        public BlobServiceBuilder(IServiceCollection serviceCollection, BlobDataStoreConfiguration config)
-        {
-            Services = EnsureArg.IsNotNull(serviceCollection, nameof(serviceCollection));
-            Configuration = EnsureArg.IsNotNull(config, nameof(config));
-        }
     }
 }
