+++ src/Microsoft.Health.Dicom.SqlServer/Registration/Experimental/SqlServiceBuilderExtensions.cs
@@ -30,14 +30,17 @@ namespace Microsoft.Extensions.DependencyInjection
 
             services.Add<SqlIndexDataStoreV1>()
                 .Scoped()
+                .AsSelf()
                 .AsImplementedInterfaces();
 
             services.Add<SqlIndexDataStoreV2>()
                 .Scoped()
+                .AsSelf()
                 .AsImplementedInterfaces();
 
             services.Add<SqlIndexDataStoreV3>()
                 .Scoped()
+                .AsSelf()
                 .AsImplementedInterfaces();
 
             services.Add<SqlStoreFactory<ISqlIndexDataStore, IIndexDataStore>>()
@@ -91,9 +94,9 @@ namespace Microsoft.Extensions.DependencyInjection
             IServiceCollection services = EnsureArg.IsNotNull(builder?.Services, nameof(builder));
 
             services.Add<SqlExtendedQueryTagStoreV1>()
-              .Scoped()
-              .AsSelf()
-              .AsImplementedInterfaces();
+                .Scoped()
+                .AsSelf()
+                .AsImplementedInterfaces();
 
             services.Add<SqlExtendedQueryTagStoreV2>()
                 .Scoped()
@@ -101,14 +104,14 @@ namespace Microsoft.Extensions.DependencyInjection
                 .AsImplementedInterfaces();
 
             services.Add<SqlExtendedQueryTagStoreV3>()
-              .Scoped()
-              .AsSelf()
-              .AsImplementedInterfaces();
+                .Scoped()
+                .AsSelf()
+                .AsImplementedInterfaces();
 
             services.Add<SqlStoreFactory<ISqlExtendedQueryTagStore, IExtendedQueryTagStore>>()
-              .Scoped()
-              .AsSelf()
-              .AsImplementedInterfaces();
+                .Scoped()
+                .AsSelf()
+                .AsImplementedInterfaces();
 
             return builder;
         }
@@ -127,6 +130,7 @@ namespace Microsoft.Extensions.DependencyInjection
 
             services.Add<BackgroundSchemaVersionResolver>()
                 .Singleton()
+                .AsSelf()
                 .AsImplementedInterfaces();
 
             return builder;
